% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_functions.R
\name{class.to.list}
\alias{class.to.list}
\title{Builds a list from class object}
\usage{
class.to.list(cl)
}
\arguments{
\item{cl}{class object}
}
\description{
Class slots are tranfered to list entries.
}
\examples{
model <- set.to.class("jumpDiffusion",
             parameter = list(theta = 0.1, phi = 0.01, gamma2 = 0.1, xi = 3))
summary(class.to.list(model))
}

