% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{dNtoTimes}
\alias{dNtoTimes}
\title{Transformation of NHPP variables to event times}
\usage{
dNtoTimes(dN, t)
}
\arguments{
\item{dN}{vector of differences of counting process}

\item{t}{times of counting process}
}
\description{
Vector of Poisson process differences are translated to a vector of event times.
}

