% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawData}
\alias{drawData}
\title{Function for simulating the data}
\usage{
drawData(t, cl, start, mw = 10, plot.series = FALSE)
}
\arguments{
\item{t}{time points}

\item{cl}{class with informations of the parameters and the type of process}

\item{start}{starting point of the process}

\item{mw}{mesh width (in the case of SDE)}

\item{plot.series}{if TRUE, data is plotted}
}
\value{
vector or matrix of simulated variables
}
\description{
Simulation of variables depend on the model.
}

