% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawSDE}
\alias{drawSDE}
\title{Function for simulating diffusion process}
\usage{
drawSDE(phi, gamma2, t, b, fODE, sigmaTilde, mw = 10,
  strictly.positive = TRUE)
}
\arguments{
\item{phi}{parameter \eqn{\phi}}

\item{gamma2}{parameter \eqn{\gamma^2}}

\item{t}{vector of time points}

\item{b}{drift function}

\item{fODE}{function for the starting point dependent on phi, for fixed y0: fODE = function(phi, t) y0}

\item{sigmaTilde}{variance function \eqn{s(\gamma, t, y) = \gamma sigmaTilde(t, y)}}

\item{mw}{mesh width to simulate the time-continuity}

\item{strictly.positive}{if TRUE, only positive values for process \eqn{Y_t}}
}
\value{
data series in t
}
\description{
Simulation of process defined by \eqn{dYt = b(\phi, t, Y_t)dt + \gamma sigmaTilde(t, Y_t)dW_t}.
}

