% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est_JD_Euler}
\alias{est_JD_Euler}
\title{Metropolis within Gibbs sampler}
\usage{
est_JD_Euler(X, N, t, n = 1000, start, b, s, h, priorRatio, Lambda,
  int = c("Weibull", "Exp"), rangeN = 2, propSd = 0.5)
}
\arguments{
\item{X}{vector of observed variables}

\item{N}{vector of Poisson process variables}

\item{t}{vector of time points}

\item{n}{length of Markov chain}

\item{start}{list of starting values, list(phi, theta, gamma2)}

\item{b}{drift function}

\item{s}{variance function}

\item{h}{jump high function}

\item{priorRatio}{list of functions for the prior ratio of MH step, if missing: non-informative}

\item{Lambda}{intensity rate function}

\item{int}{if Lambda is missing, one of "Weibull" or "Exp"}

\item{rangeN}{range for candidates for filtering N}

\item{propSd}{starting value for proposal standard deviation}
}
\value{
\item{phi}{estimator of \eqn{\phi}}
\item{gamma2}{estimator of \eqn{\gamma}2}
\item{theta}{estimator of \eqn{\theta}}
\item{xi}{estimator of \eqn{\xi}}
\item{N}{estimator of latent variable \eqn{N}, if not observed}
\item{prop}{storage of adapted proposal variances}
}
\description{
Bayesian estimation of the parameter of the jump diffusion process define by SDE
  \eqn{dXt =  b(\phi,t,Xt) dt + s(\gamma,t,Xt) dWt + h(\theta,t,Xt) dNt}.
}

