% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est_JD.R
\name{est_Merton}
\alias{est_Merton}
\title{Gibbs sampler}
\usage{
est_Merton(X, N, t, n = 1000, start, prior, Lambda, rangeN = 2,
  it.xi = 10)
}
\arguments{
\item{X}{vector of observed variables}

\item{N}{vector of Poisson process variables, optional: if missing, filtering}

\item{t}{vector of time points}

\item{n}{length of Markov chain}

\item{start}{list of starting values, list(thetaT, gamma2)}

\item{prior}{list of prior values, list(mu_phi, s_phi, mu_th, s_th, alpha, beta)}

\item{Lambda}{intensity rate function}

\item{rangeN}{range of candidates for filtering N}
}
\value{
\item{phi}{estimator of \eqn{\phi}}
\item{gamma2}{estimator of \eqn{\gamma}2}
\item{thetaT}{estimator of \eqn{log(1+\theta})}
\item{xi}{estimator of \eqn{\xi}}
\item{N}{estimator of latent variable \eqn{N}, if not observed}
}
\description{
Bayesian estimation of the parameter of the jump diffusion process
  \eqn{X_t = x_0 \exp( \phi t - \gamma2/2 t+\gamma W_t + \log(1+\theta) N_t)}.
}

