% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est_NHPP}
\alias{est_NHPP}
\title{Metropolis-Hastings sampler}
\usage{
est_NHPP(jumpTimes, Tend, start, n = 5000, int = c("Weibull", "Exp"),
  priorRatio, proposals, Lambda)
}
\arguments{
\item{jumpTimes}{vector of jump (or event) times}

\item{Tend}{last observation vector of the NHPP}

\item{start}{starting value}

\item{n}{length of Markov chain}

\item{int}{one out of "Weibull" or "Exp"}

\item{priorRatio}{function for prior ratio, if missing: non-informative}

\item{proposals}{list of functions "ratio" and "draw"}

\item{Lambda}{intensity rate function}
}
\value{
p x n dimensional matrix of posterior samples
}
\description{
Bayesian estimation of the parameter of the intensity rate.
}
\examples{
jumpTimes <- simN(seq(0, 1, by = 0.01), c(4, -1), len = 1, int = "Exp")$Times
chain <- est_NHPP(jumpTimes, 1, c(4, -1), int = "Exp")
par(mfrow = c(2, 1))
plot(chain[1,], type="l")
abline(h = 4, col = 2)
plot(chain[2,], type = "l")
abline(h = -1, col = 2)

print("acceptance rate:")
length(unique(chain[1,]))/length(chain[1,])
}

