% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est_reg_hiddenNHPP.R
\name{est_reg_hiddenNHPP}
\alias{est_reg_hiddenNHPP}
\title{Gibbs sampler}
\usage{
est_reg_hiddenNHPP(Y, N, t, fun, n = 1000, start, prior, Lambda,
  int = c("Weibull", "Exp"), rangeN = 2)
}
\arguments{
\item{Y}{vector of observation variables}

\item{N}{vector of Poisson process variables}

\item{t}{vector of time points}

\item{fun}{regression function}

\item{n}{length of Markov chain}

\item{start}{list of starting values}

\item{prior}{list of prior values}

\item{Lambda}{intensity rate function}

\item{int}{one out of "Weibull" or "Exp", if Lambda is missing}

\item{rangeN}{range for candidates of filtering N, if unobserved}
}
\value{
\item{theta}{Markov chains of \eqn{\theta}}
\item{gamma2}{Markov chains of \eqn{\gamma^2}}
\item{N}{if hidden: Markov chains of \eqn{N}}
\item{xi}{Markov chains of \eqn{\xi}}
}
\description{
Bayesian estimation of the parameter of the regression model
  \eqn{y_i = f(t_i, N_i, \theta) + \epsilon_i}.
}

