% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est_SEP}
\alias{est_SEP}
\title{Metropolis-Hastings sampler}
\usage{
est_SEP(jumpTimes, Tend, start, n = 5000, s = 200, f_xi, priorRatio,
  proposals, Lmax = 35, propSd)
}
\arguments{
\item{jumpTimes}{vector of event times}

\item{Tend}{last observation time, if missing => last event time}

\item{start}{starting value}

\item{n}{length of Markov chain}

\item{s}{stress range}

\item{f_xi}{exp(-f_xi) intensity function}

\item{priorRatio}{function(old, new) for prior ratio, if missing => noninformative}

\item{proposals}{list of functions: draw(old) and ratio(new, old)}

\item{Lmax}{maximal number of tension wires that can break}

\item{propSd}{proposal standard deviation}
}
\value{
p x n dimensional matrix of posterior samples, p length of vector xi
}
\description{
Bayesian estimation of the parameter of the self-exciting process.
}
\examples{
wt <- rexp(20, exp(- 1 + 0.25*200/(35-(0:19))))
jumpTimes <- cumsum(wt)
chain <- est_SEP(jumpTimes, start = c(1, 0.25), s = 200)
par(mfrow = c(2,1))
plot(chain[1,], type = "l")
abline(h = 1, col = 2)
plot(chain[2,], type = "l")
abline(h = 0.25, col = 2)

print("acceptance rate:")
length(unique(chain[1,]))/length(chain[1,])

# or:
s <- seq(100, 300, length = 10)
jumpTimes <- lapply(1:10, function(a){
wt <- rexp(20, exp(- 1 + 0.25*s[a]/(35-(0:19))))
cumsum(wt)
})
chain <- est_SEP(jumpTimes, start = c(1, 0.25), s = s, n = 1000)
}

