% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,Merton-method}
\alias{estimate,Merton-method}
\title{Estimation for jump diffusion process}
\usage{
\S4method{estimate}{Merton}(model.class, t, data, nMCMC)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}
}
\description{
Bayesian estimation of a stochastic process
  \eqn{Y_t = y_0 \exp( \phi t - \gamma2/2 t+\gamma W_t + \log(1+\theta) N_t)}.
}
\examples{
cl <- set.to.class("Merton", parameter = list(thetaT = 0.1, phi = 0.05, gamma2 = 0.1, xi = 10))
t <- seq(0, 1, by = 0.01)
data <- simulate(cl, t = t, y0 = 0.5, plot.series = TRUE)
est <- estimate(cl, t, data, 1000)
plot(est)
\dontrun{
est_hidden <- estimate(cl, t, data$Y, 1000)
plot(est_hidden)
}
}

