% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,NHPP-method}
\alias{estimate,NHPP-method}
\title{Estimation for Poisson process}
\usage{
\S4method{estimate}{NHPP}(model.class, t, data, nMCMC, propSd, adapt = TRUE,
  proposal = c("lognormal", "normal"))
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}

\item{propSd}{vector of proposal variances for \eqn{\xi}}

\item{adapt}{if TRUE (default), proposal variance is adapted}

\item{proposal}{"lognormal" (for positive parameters, default) or "normal"}
}
\description{
Bayesian estimation of a nonhomogeneous Poisson process.
}
\examples{
model <- set.to.class("NHPP", parameter = list(xi = c(5, 1/2)),
                   Lambda = function(t, xi) (t/xi[2])^xi[1])
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, plot.series = TRUE)
est <- estimate(model, t, data$Times, 10000)
plot(est)
##
model <- set.to.class("NHPP", parameter = list(xi = 5),
                   Lambda = function(t, xi) t*xi)
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, plot.series = TRUE)
est <- estimate(model, t, data$N, 10000, proposal = "normal")
plot(est, par.options = list(mfrow = c(1,1)))
}
\references{
Hermann, S., K. Ickstadt and C. H. Mueller (2016).
Bayesian Prediction for a Jump Diffusion Process with Application to Crack Growth in Fatigue Experiments.
SFB 823 discussion paper 30/15.
}

