% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{estimate,reg_hiddenNHPP-method}
\alias{estimate,reg_hiddenNHPP-method}
\title{Estimation for regression model dependent on Poisson process}
\usage{
\S4method{estimate}{reg_hiddenNHPP}(model.class, t, data, nMCMC)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}
}
\description{
Bayesian estimation of the parameter of the regression model
  \eqn{y_i = f(t_i, N_i, \theta) + \epsilon_i}.
}
\examples{
t <- seq(0,1, by = 0.01)
cl <- set.to.class("reg_hiddenNHPP", fun = function(t, N, theta) exp(theta[1]*t) + theta[2]*N,
                   parameter = list(theta = c(2, 2), gamma2 = 0.25, xi = c(3, 0.5)),
                   Lambda = function(t, xi) (t/xi[2])^xi[1])
data <- simulate(cl, t = t)
est <- estimate(cl, t, data, 1000)
plot(est)
\dontrun{
est_hid <- estimate(cl, t, data$Y, 1000)
plot(est_hid)
}
}

