% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,Regression-method}
\alias{estimate,Regression-method}
\title{Estimation for regression model}
\usage{
\S4method{estimate}{Regression}(model.class, t, data, nMCMC)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}
}
\description{
Bayesian estimation of the parameter of the regression model
  \eqn{y_i = f(\phi, t_i) + \epsilon_i}.
}
\examples{
t <- seq(0,1, by = 0.01)
cl <- set.to.class("Regression", fun = function(phi, t) phi[1]*t + phi[2],
                   parameter = list(phi = c(1,2), gamma2 = 0.1))
data <- simulate(cl, t = t, plot.series = TRUE)
est <- estimate(cl, t, data, 1000)
plot(est)
}

