% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estSDE}
\alias{estSDE}
\title{Bayesian estimation in mixed stochastic differential equations}
\usage{
estSDE(t, y, prior, start, bSDE, sVar, ipred = 1, cut, len = 1000,
  mod = c("Gompertz", "logistic", "Weibull", "Richards", "Paris", "Paris2"),
  propPar = 0.2)
}
\arguments{
\item{t}{vector of observation times}

\item{y}{matrix or list of the n trajectories}

\item{prior}{list of prior parameters - list(m, v, alpha.omega, beta.omega, alpha.gamma, beta.gamma)}

\item{start}{list of starting values}

\item{bSDE}{b(phi, t, x) drift function}

\item{sVar}{variance function s^2}

\item{ipred}{which of the n trajectories is the one to be predicted}

\item{cut}{the index how many of the ipred-th series are used for estimation}

\item{len}{number of iterations of the MCMC algorithm - chain length}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull, Paris, Paris2}, only used instead of bSDE}

\item{propPar}{proposal standard deviation of phi is |start$mu|*propPar}
}
\value{
\item{phi}{samples from posterior of \eqn{\phi}}
\item{mu}{samples from posterior of \eqn{\mu}}
\item{Omega}{samples from posterior of \eqn{\Omega}}
\item{gamma2}{samples from posterior of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the random effects \eqn{\phi_i} in the mixed SDE
 \eqn{dY_i(t) = b(\phi_i, t, Y_i(t))dt + \gamma s(t, Y_i(t)) dW_i(t), \phi_i~N(\mu, \Omega), i=1,...,n} and the parameters
 \eqn{\mu, \Omega, \gamma^2}.
}
\details{
Simulation from the posterior distribution of the random effect from n independent trajectories of the SDE (the Brownian motions \eqn{W1,...,Wn} are independent).
}
\examples{
mu <- 0.5; Omega <- 0.1; phi <- matrix(rnorm(100, mu, sqrt(Omega)))
cl <- set.to.class("mixedDiffusion", parameter=list(phi = phi, mu = mu, Omega = Omega, gamma2 = 0.01))
t <- seq(0, 1, by = 0.1)
Y <- simulate(cl, t = t, y0 = 0.5, plot.series = TRUE)

prior <- list(m = 1, v = 1, alpha.omega = 3, beta.omega = 2*Omega, alpha.gamma = 3, beta.gamma = 2*0.01)
start <- list(phi = phi, mu = mu, gamma2 = 0.01)
est <- estSDE(t, Y, bSDE = cl@b.fun, prior = prior, start = start, sVar = cl@sT.fun, len = 1000)
par(mfrow = c(1,2))
plot(est$mu, type = "l", ylab = expression(mu))
plot(est$gamma2, type = "l", ylab = expression(gamma^2))
}
\references{
Hermann et al. (2015)
}
\keyword{estimation}

