% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estSDE.R
\name{estSDE_single}
\alias{estSDE_single}
\title{Bayesian estimation in stochastic differential equations}
\usage{
estSDE_single(t, X, prior, start, bSDE, sVar, len = 1000)
}
\arguments{
\item{t}{vector of observation times}

\item{X}{vector of the M trajectories}

\item{prior}{list of prior parameters - list(mu, Omega, alpha, beta)}

\item{start}{list of starting values}

\item{bSDE}{drift function}

\item{sVar}{variance function}

\item{len}{number of iterations of the MCMC algorithm}
}
\value{
\item{phi}{estimator of \eqn{\phi}}
\item{gamma2}{estimator of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the parameters in the SDE
 \eqn{dY(t)= b(\phi, t, Y(t))dt + \gamma s(t, Y(t)) dW(t)}.
}
\details{
Simulation from the posterior distribution of the random effect from n independent trajectories of the SDE (the Brownian motions \eqn{W1,...,Wn} are independent).
}

