% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findCandidateArea}
\alias{findCandidateArea}
\title{Helping function}
\usage{
findCandidateArea(VFun, start = 1, pos.support = TRUE, quasi.null = 1e-05)
}
\arguments{
\item{VFun}{cumulative distribution function}

\item{start}{starting point for search}

\item{pos.support}{if TRUE: only positive support}

\item{quasi.null}{size of values to be defined as zero, default: 10^{-5}}
}
\value{
adjusted proposal standard deviation
}
\description{
Finding suitable candidate area
}

