% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPrior.R
\name{getPrior}
\alias{getPrior}
\title{Builds list of prior parameters}
\usage{
getPrior(parameter, model = c("jumpDiffusion", "Merton", "Diffusion",
  "mixedDiffusion", "hiddenDiffusion", "hiddenmixedDiffusion", "reg_hiddenNHPP",
  "NHPP", "Regression", "mixedRegression"))
}
\arguments{
\item{parameter}{list of parameters}

\item{model}{name of model}
}
\value{
list of prior values
}
\description{
Creation of list of parameters conditional on true values \eqn{\mu} and \eqn{\gamma^2}
}

