% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{mixedDiffusion-class}
\alias{mixedDiffusion-class}
\title{S4 class of model informations for hierarchical (mixed) diffusion process model}
\description{
Informations of model
\eqn{dY_t = b(\phi_j,t,Y_t)dt + \gamma \widetilde{s}(t,Y_t)dW_t, \phi_j\sim N(\mu, \Omega), Y_{t_0}=y_0(\phi, t_0)}.
}
\section{Slots}{

\describe{
\item{\code{phi}}{parameter \eqn{\phi}}

\item{\code{mu}}{parameter \eqn{\mu}}

\item{\code{Omega}}{parameter \eqn{\Omega}}

\item{\code{gamma2}}{parameter \eqn{\gamma^2}}

\item{\code{y0.fun}}{function \eqn{y_0(\phi, t)}}

\item{\code{b.fun}}{function \eqn{b(\phi,t,y)}}

\item{\code{sT.fun}}{function \eqn{\widetilde{s}(t,y)}}

\item{\code{prior}}{list of prior parameters}

\item{\code{start}}{list of starting values for the Metropolis within Gibbs sampler}
}}
\examples{
mu <- c(2, 1); Omega <- c(1, 0.04)
phi <- sapply(1:2, function(i) rnorm(21, mu[i], sqrt(Omega[i])))
parameter <- list(phi = phi, mu = mu, Omega = Omega, gamma2 = 0.01)
b.fun <- function(phi, t, y) phi[1] * y
sT.fun <- function(t, y) y
y0.fun <- function(phi, t) phi[2]
start <- parameter
prior <- list(m.mu = parameter$mu, v.mu = parameter$mu^2,
   alpha.omega = rep(3, length(parameter$mu)), beta.omega = parameter$Omega*2,
   alpha.gamma = 3, beta.gamma = parameter$gamma2*2)
model <- set.to.class("mixedDiffusion", parameter, prior, start,
  b.fun = b.fun, sT.fun = sT.fun, y0.fun = y0.fun)
}

