% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/particleGibbs.R
\name{partFiltering}
\alias{partFiltering}
\title{Estimation function}
\usage{
partFiltering(t, y, prior, start, est = c("Bayes", "MLE"), len = 1000,
  sigmaTilde, Npart = 10, y0.fun = 1, b.fun = 1, parPropPhi = 5,
  maxIt = 10)
}
\arguments{
\item{t}{vector of time points}

\item{y}{vector of observation variables}

\item{prior}{list of prior values}

\item{start}{list of starting values}

\item{est}{one out of "Bayes" or "MLE"}

\item{len}{length of Markov chain}

\item{sigmaTilde}{variance function \eqn{s(\gamma, t, y) = \gamma sigmaTilde(t, y)}}

\item{Npart}{number of particles}

\item{y0.fun}{function for starting point dependent on \eqn{\phi}}

\item{b.fun}{drift function}

\item{parPropPhi}{parameter for proposal standard deviation}

\item{maxIt}{maximal iteration of MH step of \eqn{\phi}}
}
\value{
\item{phi}{Markov chain of \eqn{\phi}}
\item{gamma2}{Markov chain of \eqn{\gamma}2}
\item{sigma2}{Markov chain of \eqn{\sigma}2}
\item{X}{filtered process}
}
\description{
Bayesian estimation of the parameter of the model \eqn{Y_i = X_{t_i} + \epsilon_i} with
  \eqn{dX_t = b(\phi,t,X_t)dt + s(\gamma,t,X_t)dW_t}.
}
\author{
Simone Hermann and Adeline Leclercq-Samson
}

