% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/particleGibbs.R
\name{partFiltering_mixed}
\alias{partFiltering_mixed}
\title{Estimation function}
\usage{
partFiltering_mixed(t, y, prior, start, len = 1000, sigmaTilde, y0.fun = 1,
  b.fun = 1, Npart = 10, parPropPhi = 5, maxIt = 10)
}
\arguments{
\item{t}{vector of time points}

\item{y}{matrix of observation variables}

\item{prior}{list of prior values}

\item{start}{list of starting values}

\item{len}{length of Markov chain}

\item{sigmaTilde}{variance function \eqn{s(\gamma, t, y) = \gamma sigmaTilde(t, y)}}

\item{y0.fun}{function for starting point dependent on \eqn{\phi}}

\item{b.fun}{drift function}

\item{Npart}{number of particles}

\item{parPropPhi}{parameter for proposal standard deviation}

\item{maxIt}{maximal iteration of MH step of \eqn{\phi}}
}
\value{
\item{phi}{Markov chain of \eqn{\phi}}
\item{mu}{Markov chain of \eqn{\mu}}
\item{Omega}{Markov chain of \eqn{\Omega}}
\item{gamma2}{Markov chain of \eqn{\gamma^2}}
\item{sigma2}{Markov chain of \eqn{\sigma^2}}
\item{X}{filtered process}
}
\description{
Bayesian estimation of the parameter of the model \eqn{Y_{ij} = X_{t_{ij}} + \epsilon_{ij}} with
  \eqn{dX_t = b(\phi_j, t, X_t)dt + s(\gamma, t, X_t)dW_t}.
}
\author{
Simone Hermann and Adeline Leclercq-Samson
}

