% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{phi_ij}
\alias{phi_ij}
\title{Helping function}
\usage{
phi_ij(phi, i, j)
}
\arguments{
\item{phi}{list with each entry a matrix}

\item{i}{row}

\item{j}{column}
}
\value{
vector of samples
}
\description{
Returns the ij th matrix entry from a list
}

