% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,est.mixedDiffusion-method}
\alias{plot,est.mixedDiffusion-method}
\title{Plot method for the Bayesian estimation class object}
\usage{
\S4method{plot}{est.mixedDiffusion}(x, newwindow = FALSE, ...)
}
\arguments{
\item{x}{est.mixedDiffusion class}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class Bayes.fit
}

