% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotQuantiles}
\alias{plotQuantiles}
\title{Plot function for credibility or prediction intervals}
\usage{
plotQuantiles(input, xlab = "", ylab = "", main = "", l = 0.025,
  u = 0.975, color = 1)
}
\arguments{
\item{input}{list or matrix of samples from posterior or predictive distribution}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{main}{an overall title for the plot}

\item{l}{lower bound}

\item{u}{upper bound}

\item{color}{color of the lines to be drawn}
}
\description{
Plots intervals.
}

