% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postOmega.R
\name{postOmega_matrix}
\alias{postOmega_matrix}
\title{Posterior}
\usage{
postOmega_matrix(R, phi, mu)
}
\arguments{
\item{R}{prior matrix of wishart distribution}

\item{phi}{matrix of random effects}

\item{mu}{mean of random effects}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\Omega}
}

