% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pred_SEP}
\alias{pred_SEP}
\title{Prediction of SEP}
\usage{
pred_SEP(xi, s, J = 1, f_xi_mat, len = 100, end = 20, grid = 0.01,
  Lmax = 35)
}
\arguments{
\item{xi}{Markov chains}

\item{s}{stress}

\item{J}{index}

\item{f_xi_mat}{matrix-vise function}

\item{len}{length of samples}

\item{end}{how many jumps}

\item{grid}{fineness degree}

\item{Lmax}{maximal number of tension wires that can break}
}
\description{
Bayesian prediction of the self-exciting process.
}
\examples{
s <- seq(100, 300, length = 10)
f_xi <- function(y, xi) xi[1] - xi[2]*y
jumpTimes <- lapply(1:10, function(a){
wt <- rexp(20, exp(- f_xi(s[a]/(35-(0:19)), c(1, 0.25))))
cumsum(wt)
})
\dontrun{
chain <- est_SEP(jumpTimes, start = c(1, 0.25), s = s, n = 11000)
pred <- pred_SEP(chain[seq(1001, 11000, by = 10),], s = s[1], J = 1, f_xi_mat = function(y, xi) xi[,1] - xi[,2]*y)
qu <- apply(pred, 2, quantile, c(0.025, 0.975))
plot(jumpTimes[[1]], ylim = range(c(jumpTimes, range(qu))))
lines(qu[1,])
lines(qu[2,])
}
}

