% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction.intervals}
\alias{prediction.intervals}
\title{Bayesian prediction interval function}
\usage{
prediction.intervals(samples, Fun, x0, level = 0.05, candArea, grid = 0.001)
}
\arguments{
\item{samples}{posterior samples}

\item{Fun}{cumulative distribution function}

\item{x0}{starting point}

\item{level}{level of prediction intervals}

\item{candArea}{candidate area}

\item{grid}{fineness degree for the binary search}
}
\value{
prediction interval
}
\description{
Calculation of quantiles \code{level}/2 and 1-\code{level}/2 from predictive distribution based on distribution function \code{Fun(x, x0, samples)}.
\code{samples} should contain samples from the posterior distribution of the parameters.
}

