% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predPhi}
\alias{predPhi}
\title{Prediction Of The Random Effects In Mixed Stochastic Differential Equations}
\usage{
predPhi(samples, cand)
}
\arguments{
\item{samples}{output of estSDE or estREg}

\item{cand}{candidates for phi (matrix with p columns)}
}
\value{
matrix phi
}
\description{
Prediction of the random effects \eqn{\phi ~ N(\mu, \Omega)}
}

