% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{scoreRule}
\alias{scoreRule}
\title{Calculation of interval score}
\usage{
scoreRule(l, u, x, alpha = 0.05)
}
\arguments{
\item{l}{lower bound}

\item{u}{upper bound}

\item{x}{true value}

\item{alpha}{level}
}
\value{
interval score
}
\description{
Scoring rule of Raftery and Gneiting (??).
}

