% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_JD_Euler}
\alias{sim_JD_Euler}
\title{Simulation of Jump diffusion process}
\usage{
sim_JD_Euler(t, phi, theta, gamma2, b.fun, s.fun, h.fun, start, N, mw = 1)
}
\arguments{
\item{t}{vector of times}

\item{phi}{parameter \eqn{\phi}}

\item{theta}{parameter \eqn{\theta}}

\item{gamma2}{parameter \eqn{\gamma^2}}

\item{b.fun}{drift function}

\item{s.fun}{variance function}

\item{h.fun}{jump high function}

\item{start}{starting point \eqn{y_0}}

\item{N}{Poisson process variables in t}

\item{mw}{mesh width for finer Euler approximation}
}
\value{
matrix or vector
}
\description{
Simulation of Jump diffusion process.
}

