% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_reg_hiddenNHPP}
\alias{sim_reg_hiddenNHPP}
\title{Simulation of regression model including the NHPP}
\usage{
sim_reg_hiddenNHPP(t, N, fun, theta, gamma2)
}
\arguments{
\item{t}{vector of times}

\item{N}{vector of Poisson process}

\item{fun}{regression function}

\item{theta}{parameter \eqn{\theta}}

\item{gamma2}{parameter\eqn{\gamma^2}}
}
\value{
matrix or vector
}
\description{
Simulation.
}

