% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simN}
\alias{simN}
\title{Simulation of counting process}
\usage{
simN(t, xi, len, start = c(0, 0), Lambda, int = c("Weibull", "Exp"))
}
\arguments{
\item{t}{vector of times}

\item{xi}{parameter vector \eqn{\xi}}

\item{len}{number of samples to be drawn}

\item{start}{vector: start[1] starting point time, start[2] starting point for Poisson process}

\item{Lambda}{intensity rate function}

\item{int}{if no Lambda: one out of "Weibull" or "Exp" for intensity function}
}
\value{
\item{N}{Poisson process}
\item{Times}{event times}
}
\description{
Simulation of counting process and event times.
}

