% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simY}
\alias{simY}
\title{Simulation of Jump diffusion process}
\usage{
simY(t, phi, thetaT, gamma2, start, N)
}
\arguments{
\item{t}{vector of times}

\item{phi}{parameter \eqn{\phi}}

\item{thetaT}{parameter \eqn{\widetilde{\theta}}}

\item{gamma2}{parameter \eqn{\gamma^2}}

\item{start}{starting point ofprocess \eqn{y_0}}

\item{N}{Poisson process variables in t}
}
\value{
matrix or vector
}
\description{
Simulation of Jump diffusion process.
}

