% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SMC}
\alias{SMC}
\title{SMC}
\usage{
SMC(phi, gamma2, sigma2, Npart, times, y, b.fun, y0.fun, sigmaTilde,
  conditional = TRUE, X.cond, B_fixed)
}
\arguments{
\item{phi}{parameter \eqn{\phi}}

\item{gamma2}{parameter \eqn{\gamma^2}}

\item{sigma2}{parameter \eqn{\sigma^2}}

\item{Npart}{number of particle}

\item{times}{vector of time points}

\item{y}{observation vector}

\item{b.fun}{drift function}

\item{y0.fun}{function for starting point dependent on \eqn{\phi}}

\item{sigmaTilde}{variance function}

\item{conditional}{logical(1), if TRUE conditional SMC}

\item{X.cond}{if conditional = TRUE, the series to be hold}

\item{B_fixed}{ancestral lineage of X.cond}
}
\description{
Sequential Monte Carlo step - conditional and unconditional.
}
\author{
Simone Hermann and Adeline Leclercq-Samson
}

