% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TimestoN}
\alias{TimestoN}
\title{Transformation of event times to NHPP}
\usage{
TimestoN(times, t)
}
\arguments{
\item{times}{vector of event times}

\item{t}{times of counting process}
}
\description{
Transformation of vector of event times to counting process.
}

