% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{ad.propSd}
\alias{ad.propSd}
\title{Helping function}
\usage{
ad.propSd(chain, propSd, iteration, lower = 0.3, upper = 0.6,
  delta.n = function(n) min(0.05, 1/sqrt(n)))
}
\arguments{
\item{chain}{Markov chain}

\item{propSd}{current proposal standard deviation}

\item{iteration}{current iteration (batch)}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{delta.n}{function of batch number}
}
\value{
adjusted proposal standard deviation
}
\description{
Adaptive MCMC
}

