% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagnostic}
\alias{diagnostic}
\title{Calcucation of burn-in phase and thin rate}
\usage{
diagnostic(chain, dependence = 0.8, m = 10)
}
\arguments{
\item{chain}{vector of Markov chain samples}

\item{dependence}{allowed dependence for the chain}

\item{m}{number of blocks}
}
\description{
Proposal for burn-in and thin rate
}

