% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dNtoTimes}
\alias{dNtoTimes}
\title{Transformation of counting process to vector of event times}
\usage{
dNtoTimes(dN, t)
}
\arguments{
\item{dN}{vector of differences of counting process}

\item{t}{times of counting process}
}
\value{
vector of event times
}
\description{
Transformation of vector of counting process to event times.
}
\examples{
t <- seq(0, 1, by = 0.01)
process <- simN(t, c(5, 0.5), len = 1)$N
times <- dNtoTimes(diff(process), t)
}

