% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est_NHPP}
\alias{est_NHPP}
\title{Metropolis-Hastings sampler}
\usage{
est_NHPP(jumpTimes, Tend, start, n = 5000, int = c("Weibull", "Exp"),
  priorRatio, proposal = c("lognormal", "normal"), Lambda)
}
\arguments{
\item{jumpTimes}{vector of jump (or event) times}

\item{Tend}{last observation vector of the NHPP}

\item{start}{starting value}

\item{n}{length of Markov chain}

\item{int}{one out of "Weibull" or "Exp"}

\item{priorRatio}{function for prior ratio, if missing: non-informative}

\item{proposal}{"lognormal" (for positive parameters, default) or "normal"}

\item{Lambda}{intensity rate function}
}
\value{
p x n dimensional matrix of posterior samples
}
\description{
Bayesian estimation of the parameter of the intensity rate.
}
\examples{
# exponential
Lambda <- function(t, xi){
xi[2]*exp(xi[1]*t)-xi[2]
}
jumpTimes <- simN(seq(0, 1, by = 0.001), c(4, 0.5), len = 1, Lambda = Lambda)$Times
chain <- est_NHPP(jumpTimes, 1, c(4, 0.5), Lambda = Lambda)
# chain <- est_NHPP(jumpTimes, 1, c(4, 0.5), int = "Exp")
plot(chain[1,], type="l"); abline(h = 4)
plot(chain[2,], type="l"); abline(h = 0.5)

# weibull
Lambda <- function(t, xi){
(t/xi[2])^xi[1]
}
jumpTimes <- simN(seq(0, 1, by = 0.001), c(3, 0.5), len = 1, Lambda = Lambda)$Times
chain <- est_NHPP(jumpTimes, 1, c(3, 0.5), Lambda = Lambda)
# chain <- est_NHPP(jumpTimes, 1, c(3, 0.5), int = "Weibull")
plot(chain[1,], type="l"); abline(h = 3)
plot(chain[2,], type="l"); abline(h = 0.5)
}

