% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,hiddenDiffusion-method}
\alias{estimate,hiddenDiffusion-method}
\title{Estimation for noisy / hidden diffusion process}
\usage{
\S4method{estimate}{hiddenDiffusion}(model.class, t, data, nMCMC, Npart = 100)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}

\item{Npart}{number of particles in the particle Gibbs sampler}
}
\description{
Bayesian estimation of the model,
  \eqn{Z_i = Y_{t_i} + \epsilon_i, dY_t = b(\phi,t,Y_t)dt + s(\gamma,t,Y_t)dW_t}.
}
\examples{
cl <- set.to.class("hiddenDiffusion", y0.fun = function(phi, t) 0.5,
             parameter = list(phi = 5, gamma2 = 1, sigma2 = 0.1))
t <- seq(0, 1, by = 0.01)
data <- simulate(cl, t = t, plot.series = TRUE)
est <- estimate(cl, t, data$Z, 1000)
plot(est)
# OU
b.fun <- function(phi, t, y) phi[1]-phi[2]*y
cl <- set.to.class("hiddenDiffusion", y0.fun = function(phi, t) 0.5,
               parameter = list(phi = c(10, 5), gamma2 = 1, sigma2 = 0.1),
               b.fun = b.fun, sT.fun = function(t, x) 1)
t <- seq(0, 1, by = 0.01)
data <- simulate(cl, t = t, plot.series = TRUE)
est <- estimate(cl, t, data$Z, 1000)
plot(est)
}

