% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,hiddenmixedDiffusion-method}
\alias{estimate,hiddenmixedDiffusion-method}
\title{Estimation for noisy/hidden mixed diffusion process}
\usage{
\S4method{estimate}{hiddenmixedDiffusion}(model.class, t, data, nMCMC, propSd,
  adapt = TRUE, Npart = 100, proposal = c("normal", "lognormal"))
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}

\item{propSd}{vector of proposal variances for \eqn{\phi}}

\item{adapt}{if TRUE (default), proposal variance is adapted}

\item{Npart}{number of particles in the particle Gibbs sampler}

\item{proposal}{proposal density "normal" (default) or "lognormal" (for positive parameters)}
}
\description{
Bayesian estimation of a stochastic process
  \eqn{Z_{ij} = Y_{t_{ij}} + \epsilon_{ij}, dY_t = b(\phi_j,t,Y_t)dt + s(\gamma,t,Y_t)dW_t, \phi_j~N(\mu, \Omega)}.
}
\examples{
mu <- c(5, 1); Omega <- c(0.9, 0.04)
phi <- cbind(rnorm(21, mu[1], sqrt(Omega[1])), rnorm(21, mu[2], sqrt(Omega[2])))
y0.fun <- function(phi, t) phi[2]
model <- set.to.class("hiddenmixedDiffusion", y0.fun = y0.fun,
                 b.fun = function(phi, t, y) phi[1],
                 parameter = list(phi = phi, mu = mu, Omega = Omega, gamma2 = 1, sigma2 = 0.01))
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, plot.series = TRUE)
\dontrun{
est <- estimate(model, t, data$Z[1:20,], 2000)
plot(est)
}
}
\references{
Andrieu, C., A. Doucet and R. Holenstein (2010). Particle Markov Chain Monte Carlo Methods.
Journal of the Royal Statistical Society B 72, pp. 269-342.
}

