% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\docType{methods}
\name{estimate,jumpDiffusion-method}
\alias{estimate,jumpDiffusion-method}
\title{Estimation for jump diffusion process}
\usage{
\S4method{estimate}{jumpDiffusion}(model.class, t, data, nMCMC, propSd,
  adapt = TRUE)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}

\item{propSd}{vector of proposal variances for \eqn{(\phi, \theta, \gamma^2, \xi)}}

\item{adapt}{if TRUE (default), proposal variance is adapted}
}
\description{
Bayesian estimation of a stochastic process
  \eqn{dY_t = b(\phi,t,Y_t)dt + s(\gamma,t,Y_t)dW_t + h(\eta,t,Y_t)dN_t}.
}
\examples{
# non-informative
model <- set.to.class("jumpDiffusion", Lambda = function(t, xi) (t/xi[2])^xi[1],
               parameter = list(theta = 0.1, phi = 0.05, gamma2 = 0.1, xi = c(3, 1/4)))
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, y0 = 0.5, plot.series = TRUE)
est <- estimate(model, t, data, 1000)
plot(est)
# informative
model <- set.to.class("jumpDiffusion", Lambda = function(t, xi) (t/xi[2])^xi[1],
   parameter = list(theta = 0.1, phi = 0.05, gamma2 = 0.1, xi = c(3, 1/4)),
   priorRatio = list(phi = function(new, old) dnorm(new, 0.05, 0.01)/dnorm(old, 0.05, 0.01),
               theta = function(new, old) dgamma(1/new, 10, 0.1*9)/dgamma(1/old, 10, 0.1*9),
             gamma2 = function(new, old) dgamma(1/new, 10, 0.1*9)/dgamma(1/old, 10, 0.1*9),
xi = function(new, old) prod(dnorm(new, c(3, 1/4), c(1,1))/dnorm(new, c(3, 1/4), c(1,1)))))
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, y0 = 0.5, plot.series = TRUE)
est <- estimate(model, t, data, 1000)
plot(est)
\dontrun{
est_hidden <- estimate(model, t, data$Y, 1000)
plot(est_hidden)
}
}

