% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{estimate,NHPP-method}
\alias{estimate,NHPP-method}
\title{Estimation for Poisson process}
\usage{
\S4method{estimate}{NHPP}(model.class, t, data, nMCMC)
}
\arguments{
\item{model.class}{class of the respective model including all required information, see function set.to.class}

\item{t}{vector of time points}

\item{data}{vector or list or matrix of observation variables}

\item{nMCMC}{length of Markov chain}
}
\description{
Bayesian estimation of a nonhomogeneous Poisson process.
}
\examples{
model <- set.to.class("NHPP", parameter = list(xi = c(5, 1/2)),
                   Lambda = function(t, xi) (t/xi[2])^xi[1])
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, plot.series = TRUE)
est_NHPP <- estimate(model, t, data$Times, 10000)
plot(est_NHPP)
}

