% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimation}
\usage{
estimate(model.class, ...)
}
\arguments{
\item{model.class}{class}

\item{...}{parameters dependent on the model class}
}
\description{
Method for the S4 classes
}

