% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estSDE.R
\name{estSDE}
\alias{estSDE}
\title{Bayesian estimation in mixed stochastic differential equations}
\usage{
estSDE(t, y, prior, start, y0.fun, bSDE, sVar, ipred = 1, cut, len = 1000,
  mod = c("Gompertz", "logistic", "Weibull", "Richards", "Paris", "Paris2"),
  propPar = 0.2)
}
\arguments{
\item{t}{vector of observation times}

\item{y}{matrix or list of the n trajectories}

\item{prior}{list of prior parameters - list(m, v, alpha.omega, beta.omega, alpha.gamma, beta.gamma)}

\item{start}{list of starting values}

\item{y0.fun}{\eqn{y_0(\phi, t_1)} function}

\item{bSDE}{b(phi, t, x) drift function}

\item{sVar}{variance function s^2}

\item{ipred}{which of the n trajectories is the one to be predicted}

\item{cut}{the index how many of the ipred-th series are used for estimation}

\item{len}{number of iterations of the MCMC algorithm - chain length}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull, Paris, Paris2}, only used instead of bSDE}

\item{propPar}{proposal standard deviation of phi is |start$mu|*propPar}
}
\value{
\item{phi}{samples from posterior of \eqn{\phi}}
\item{mu}{samples from posterior of \eqn{\mu}}
\item{Omega}{samples from posterior of \eqn{\Omega}}
\item{gamma2}{samples from posterior of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the random effects \eqn{\phi_i} in the mixed SDE
 \eqn{dY_i(t) = b(\phi_i, t, Y_i(t))dt + \gamma s(t, Y_i(t)) dW_i(t), \phi_i~N(\mu, \Omega), i=1,...,n} and the parameters
 \eqn{\mu, \Omega, \gamma^2}.
}
\details{
Simulation from the posterior distribution of the random effect from n independent trajectories of the SDE (the Brownian motions \eqn{W1,...,Wn} are independent).
}
\references{
Hermann et al. (2015)
}
\keyword{estimation}

