% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{jumpDiffusion-class}
\alias{jumpDiffusion-class}
\title{S4 class for the jump diffusion process}
\description{
S4 class for the jump diffusion process
}
\section{Slots}{

\describe{
\item{\code{theta}}{}

\item{\code{phi}}{...}
}}

