% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,est.Regression-method}
\alias{plot,est.Regression-method}
\title{Plot method for the Bayesian estimation class object}
\usage{
\S4method{plot}{est.Regression}(x, par.options, style = c("chains", "acf",
  "density"), par2plot, reduced = FALSE, thinning, burnIn,
  priorMeans = TRUE, col.priorMean = 2, lty.priorMean = 1,
  newwindow = FALSE, ...)
}
\arguments{
\item{x}{est.Regression class}

\item{par.options}{list of options for function par()}

\item{style}{one out of "chains", "acf", "density"}

\item{par2plot}{logical vector, which parameters to be plotted, order: \eqn{(\phi, \gamma^2)}}

\item{reduced}{logical (1), if TRUE, the chains are thinned and burn-in phase is dropped}

\item{thinning}{thinning rate, if missing, the proposed one by the estimation procedure is taken}

\item{burnIn}{burn-in phase, if missing, the proposed one by the estimation procedure is taken}

\item{priorMeans}{logical(1), if TRUE (default), prior means are marked with a line}

\item{col.priorMean}{color of the prior mean line, default 2}

\item{lty.priorMean}{linetype of the prior mean line, default 1}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class est.Regression
}
\examples{
model <- set.to.class("Regression", fun = function(phi, t) phi[1]*t + phi[2],
    parameter = list(phi = c(1, 2), gamma2 = 0.1))
data <- simulate(model, t = seq(0, 1, by = 0.01), plot.series = TRUE)
est <- estimate(model, t = seq(0, 1, by = 0.01), data, 1000)  # nMCMC small for example
plot(est)
plot(est, burnIn = 100, thinning = 2, reduced = TRUE)
plot(est, par.options = list(mar = c(5, 4.5, 4, 2) + 0.1, mfrow = c(3,1)), xlab = "iteration")
plot(est, style = "acf", main = "", par2plot = c(TRUE, TRUE, FALSE))
plot(est, style = "density", lwd = 2, priorMean = FALSE)
plot(est, style = "density", col.priorMean = 1, lty.priorMean = 2, main = "posterior")
plot(est, style = "acf", par.options = list(), main = "", par2plot = c(FALSE, FALSE, TRUE))
\dontrun{
plot(est, priorMeans = FALSE, newwindow = TRUE)
}
}

