% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmu.R
\name{postmu}
\alias{postmu}
\title{Posterior for mu}
\usage{
postmu(phi, m, v, Omega)
}
\arguments{
\item{phi}{matrix of random effects}

\item{m}{prior mean}

\item{v}{prior variance}

\item{Omega}{variance of the random effects}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\mu}
}

