% Generated by roxygen2 (4.0.1): do not edit by hand
\name{postOmega}
\alias{postOmega}
\title{Posterior}
\usage{
postOmega(alpha, beta, phi, mu)
}
\arguments{
\item{alpha}{vector of prior variables}

\item{beta}{vector of prior variables}

\item{phi}{matrix of random effects}

\item{mu}{mean of random effects}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\Omega}
}

