% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pred.base}
\alias{pred.base}
\title{Prediction function}
\usage{
pred.base(samples, VFun, dens, len = 100, x0, method = c("vector", "free"),
  pred.alg = c("Distribution", "Trajectory"), sampling.alg = c("RejSamp",
  "BinSearch"), candArea, grid = 0.001)
}
\arguments{
\item{samples}{MCMC samples}

\item{VFun}{cumulative distribution function}

\item{dens}{density function}

\item{len}{number of samples to be drawn}

\item{x0}{starting point}

\item{method}{vectorial ("vector") or not ("free")}

\item{pred.alg}{prediction algorithm, "Distribution" or "Trajectory"}

\item{sampling.alg}{sampling algorithm, rejection sampling ("RejSamp") or inversion method ("BinSearch")}

\item{candArea}{candidate area}

\item{grid}{fineness degree}
}
\value{
vector of samples from prediction
}
\description{
Bayesian prediction of the parameter of a stochastic process
  \eqn{dY_t = b(\phi, t, Y_t)dt + s(\gamma, t, Y_t)dW_t + h(\eta, t, Y_t)dN_t}.
}

