% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction.intervals}
\alias{prediction.intervals}
\title{Prediction interval function}
\usage{
prediction.intervals(samples, Fun, x0, level = 0.05, candArea, grid = 0.001)
}
\arguments{
\item{samples}{MCMC samples}

\item{Fun}{cumulative distribution function}

\item{x0}{starting point}

\item{level}{level of prediction intervals}

\item{candArea}{candidate area}

\item{grid}{fineness degree}
}
\value{
prediction interval
}
\description{
Bayesian prediction of the parameter of a stochastic process
  \eqn{dY_t = b(phi, t, Y_t)dt + s(\gamma, t, Y_t)dW_t + h(\eta, t, Y_t)dN_t}.
}

