% Generated by roxygen2 (4.0.1): do not edit by hand
\name{proposal}
\alias{proposal}
\title{Sampling from lognormal proposal density}
\usage{
proposal(parOld, propSd)
}
\arguments{
\item{parOld}{the parameter from the last iteration step}

\item{propSd}{proposal standard deviation}
}
\description{
Drawing of one sample from the lognormal distribution with mean \code{parOld} and standard deviation \code{propSd}. Used in Metropolis Hastings algorithms.
}
\examples{
plot(replicate(100, proposal(1, 0.1)), type = "l")
}

