% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{proposal}
\alias{proposal}
\title{Sampling from proposal density for strictly positive parameters}
\usage{
proposal(parOld, propSd)
}
\arguments{
\item{parOld}{the parameter from the last iteration step}

\item{propSd}{proposal standard deviation}
}
\value{
candidate for MH ratio
}
\description{
Used in Metropolis Hastings algorithms.
}
\examples{
proposal(1, 0.01)
}

