% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{proposalRatio}
\alias{proposalRatio}
\title{Proposal ratio of lognormal proposal density}
\usage{
proposalRatio(parOld, parNew, propSd)
}
\arguments{
\item{parOld}{the parameter from the last iteration step}

\item{parNew}{drawn candidate}

\item{propSd}{proposal standard deviation}
}
\description{
Calculation of proposal ratio, see also \code{\link{proposal}}.
}
\examples{
cand <- proposal(1, 0.01)
proposalRatio(1, cand, 0.01)
}

