% Generated by roxygen2 (4.0.1): do not edit by hand
\name{proposalRatio}
\alias{proposalRatio}
\title{Sampling from proposal density for strictly positive parameters}
\usage{
proposalRatio(parOld, parNew, propSd)
}
\arguments{
\item{parOld}{the parameter from the last iteration step}

\item{parNew}{drawn candidate}

\item{propSd}{proposal standard deviation}
}
\value{
MH ratio
}
\description{
Used in Metropolis Hastings algorithms.
}
\examples{
cand <- proposal(1, 0.01)
proposalRatio(1, cand, 0.01)
}

