% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_functions.R
\name{set.to.class}
\alias{set.to.class}
\title{Builds classes}
\usage{
set.to.class(class.name = c("jumpDiffusion", "Merton", "Diffusion",
  "mixedDiffusion", "hiddenDiffusion", "hiddenmixedDiffusion", "reg_hiddenNHPP",
  "NHPP", "Regression", "mixedRegression"), parameter, prior, start, b.fun,
  s.fun, h.fun, sT.fun, y0.fun, fun, Lambda, priorRatio)
}
\arguments{
\item{class.name}{name of model class}

\item{parameter}{list of parameter values}

\item{prior}{optional list of prior parameters}

\item{start}{optional list of starting values}

\item{b.fun}{drift function b}

\item{s.fun}{variance function s}

\item{h.fun}{jump high function h}

\item{sT.fun}{variance function \eqn{\widetilde{s}}}

\item{y0.fun}{function for the starting point, if dependent on parameter}

\item{fun}{regression function}

\item{Lambda}{intensity rate of Poisson process}

\item{priorRatio}{list of functions for prior ratios, only for jumpDiffusion, is missing: non-informative estimation}
}
\value{
class
}
\description{
Defines classes
}
\examples{
set.to.class("jumpDiffusion")
cl_jd <- set.to.class("jumpDiffusion", parameter = list(theta = 0.1, phi = 0.01, gamma2 = 0.1, xi = 3))
summary(class.to.list(cl_jd))
}

