% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simulate,hiddenDiffusion-method}
\alias{simulate,hiddenDiffusion-method}
\title{Simulation of diffusion process}
\usage{
\S4method{simulate}{hiddenDiffusion}(object, nsim = 1, seed = NULL, t,
  mw = 10, plot.series = TRUE)
}
\arguments{
\item{object}{class object of parameters: "hiddenDiffusion"}

\item{nsim}{number of response vectors to simulate. Defaults to 1}

\item{seed}{optional: seed number for random number generator}

\item{t}{vector of time points to make predictions for}

\item{mw}{mesh width for finer Euler approximation}

\item{plot.series}{logical(1), if TRUE, simulated series are depicted grafically}
}
\description{
Simulation of a hidden stochastic process
  \eqn{Z_i = Y_{t_i} + \epsilon_i, dY_t = b(\phi,t,Y_t)dt + s(\gamma,t,Y_t)dW_t}.
}
\examples{
model <- set.to.class("hiddenDiffusion", parameter = list(phi = 0.5, gamma2 = 0.01, sigma2 = 0.1))
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, plot.series = TRUE)
}

