% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simulate,mixedDiffusion-method}
\alias{simulate,mixedDiffusion-method}
\title{Simulation of diffusion process}
\usage{
\S4method{simulate}{mixedDiffusion}(object, nsim = 1, seed = NULL, t, y0,
  mw = 10, plot.series = TRUE)
}
\arguments{
\item{object}{class object of parameters: "mixedDiffusion"}

\item{nsim}{number of response vectors to simulate. Defaults to 1}

\item{seed}{optional: seed number for random number generator}

\item{t}{vector of time points to make predictions for}

\item{y0}{starting point of the process}

\item{mw}{mesh width for finer Euler approximation}

\item{plot.series}{logical(1), if TRUE, simulated series are depicted grafically}
}
\description{
Simulation of a stochastic process
  \eqn{dY_t = b(\phi_j,t,Y_t)dt + s(\gamma,t,Y_t)dW_t, \phi_j~N(\mu, \Omega)}.
}
\examples{
mu <- 2; Omega <- 0.4; phi <- matrix(rnorm(21, mu, sqrt(Omega)))
model <- set.to.class("mixedDiffusion", parameter = list(phi = phi, mu = mu, Omega = Omega, gamma2 = 0.1), b.fun = function(phi, t, x) phi*x, sT.fun = function(t, x) x)
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t, y0 = 0.5, plot.series = TRUE)
}

