% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simulate,NHPP-method}
\alias{simulate,NHPP-method}
\title{Simulation of Poisson process}
\usage{
\S4method{simulate}{NHPP}(object, nsim = 1, seed = NULL, t,
  plot.series = TRUE)
}
\arguments{
\item{object}{class object of parameters: "NHPP"}

\item{nsim}{number of response vectors to simulate. Defaults to 1}

\item{seed}{optional: seed number for random number generator}

\item{t}{vector of time points to make predictions for}

\item{plot.series}{logical(1), if TRUE, simulated series are depicted grafically}
}
\description{
Simulation of non-homoegenous Poisson process.
}
\examples{
model <- set.to.class("NHPP", parameter = list(xi = c(5, 1/2)),
                      Lambda = function(t, xi) (t/xi[2])^xi[1])
t <- seq(0, 1, by = 0.01)
data <- simulate(model, t = t)
}

