% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{TimestoN}
\alias{TimestoN}
\title{Transformation of vector of event times to counting process}
\usage{
TimestoN(times, t)
}
\arguments{
\item{times}{vector of event times}

\item{t}{times of counting process}
}
\value{
vector of counting process observations in t
}
\description{
Transformation of vector of event times to counting process.
}
\examples{
t <- seq(0, 1, by = 0.01)
times <- simN(t, c(5, 0.5), len = 1)$Times
process <- TimestoN(times, t)
}

